Rename:common layout 폴더 이름 수정
