+++ app/(pages)/(auth)/register/success/page.tsx
@@ -1,5 +1,5 @@
 import Success from "@common/assets/images/register_success.svg";
-import LinkLayout from "@common/components/layout/LinkLayout/LinkLayout";
+import LinkLayout from "@common/components/layout/Link/LinkLayout";
 import Button from "@common/components/ui/buttons/Button/Button";
 
 export default function RegisterSuccessPage() {
