+++ app/common/components/layout/Slider/SliderLayout.tsx
