+++ app/feature/curation/components/CurationDetail/molecules/CurationDetailInfoCard.tsx
@@ -3,9 +3,9 @@ import ScrapFill from "@common/assets/icons/scrap/ScrapFill";
 import Image from "next/image";
 import { forwardRef } from "react";
 import { CurationPlaceProps } from "@feature/curation/type";
-import LinkLayout from "@common/components/layout/LinkLayout/LinkLayout";
+import LinkLayout from "@common/components/layout/Link/LinkLayout";
 import PlaceInfoCardBottom from "@feature/place/components/PlaceInfo/molecules/PlaceInfoCardBottom";
-import SliderLayout from "@common/components/layout/SliderLayout/SliderLayout";
+import SliderLayout from "@common/components/layout/Slider/SliderLayout";
 import ScrapLine from "@common/assets/icons/scrap/ScrapLine";
 import revalidateScrapSpace from "@feature/place/actions/revalidateScrapSpace";
 import revalidateMyCuration from "@feature/curation/actions/revalidateMyCuration";
